+++ custom_components/elco_aerotop/number.py
@@ -6,6 +6,7 @@ from homeassistant.components.number import NumberEntity, NumberMode
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import UnitOfTemperature
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.entity import EntityCategory
 
 from .coordinator import ElcoDataUpdateCoordinator
 from .entity import ElcoAerotopEntity
@@ -15,6 +16,7 @@ from .models import NumericVariable
 class ElcoTemperatureNumber(ElcoAerotopEntity, NumberEntity):
     _attr_native_unit_of_measurement = UnitOfTemperature.CELSIUS
     _attr_mode = NumberMode.BOX
+    _attr_entity_category = EntityCategory.CONFIG
 
     def __init__(
         self,